Renamed fn_asm_traits.rs to asm_traits.rs and exported it
