+++ src/rust/iced-x86/src/code_asm/fn_asm_pub.rs
@@ -3,7 +3,7 @@
 
 // ⚠️This file was generated by GENERATOR!🦹‍♂️
 
-use crate::code_asm::fn_asm_traits::*;
+use crate::code_asm::asm_traits::*;
 use crate::code_asm::CodeAssembler;
 use crate::IcedError;
